+++ Dev/BSupport/DateTime/date_time.h
@@ -16,6 +16,6 @@ void date_time_set(uint8_t day, uint8_t hour, uint8_t minute, uint8_t second);
 
 void date_time_update(void);
 
-void date_time_get(uint32_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
+void date_time_get(uint8_t *days, uint8_t *hours, uint8_t *minutes, uint8_t *seconds);
 
 #endif /* M4_UTILS_DATETIME_DATE_TIME_H_ */
